requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
djangorestframework==3.12.1
djangorestframework-stubs==1.2.0
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
uplink==0.9.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
